test: 按职责拆分 Todo Agent 集成测试
